% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu_c2}
\alias{simu_c2}
\title{A Simulated Data Set under Left Truncation and Right Censoring where Censoring is always after Truncation}
\format{
A data frame with 500 rows and 5 variables.
\describe{
  \item{X}{the censored event time X = min(T,C)}
  \item{Q}{the left truncation time}
  \item{delta}{the event indicator}
  \item{Z1}{a continous covariate}
  \item{Z2}{a binary covariate}
}
}
\source{
Simulated
}
\usage{
data(simu_c2)
}
\description{
A simulated data set under left truncation and right censoring where censoring is always after left truncation.
The left truncation time and the event time are dependent via two covariates `Z1' and `Z2'.
Under the data generating mechanism,
the conditional distributions of the event time given covariates follows a Cox model in the full data,
and the conditional distribution of the left truncation time given covariates follows a Cox model on the reversed time scale  in the full data.
The truncation rate is 29.5\%; the censoring rate is 27.1\%;
and the truth \eqn{P^*(T>3) = 0.576547}.
}
\examples{
data(simu_c2)
}
\keyword{datasets}
